Fix 404 frfr (#1084)
* add back root layout and refactor to return <html>

* refactor layout and provider hierarchy

* delete catchAll... bye!

* move navigation into the [locale]/layout

* add navigation directly to not-found component

* get locale working with 404

* cleanup

* remove test zz locale

* wrap navigation with i18n client provider

* yank nav and cleanup 404

* remove unused import

* make the navigation bar work as a server component
this allows the 404 to have a nav
as well as a single nav bar that works with i18n

* fix failing build - setStaticParamsLocale in not-found

* force dynamic on not-found 404 page

* chore: remove import + move nav into a dir

---------
