+++ apps/web/src/locales/en.ts
@@ -2,4 +2,7 @@ export default {
   'landing.hero.description':
     'Connect, collaborate, and grow with a community of TypeScript developers. Elevate your skills through interactive coding challenges, discussions, and knowledge sharing',
   'landing.hero.buttons.waitlist': 'Join the waitlist',
+  '404.message': 'Page not found, sus',
+  'navigation.explore': 'Explore',
+  'navigation.tracks': 'Tracks',
 } as const;
